requirements.txt
Flask==3.0.1
torch==2.4.0
torchvision==0.19.0
Pillow==7.0.0